+++ catalyst/callbacks/metrics/auc.py
@@ -21,7 +21,7 @@ class AUCCallback(LoaderMetricCallback):
         output_key: str = "logits",
         prefix: str = "auc",
         activation: str = "Sigmoid",
-        log_per_class: bool = False,
+        per_class: bool = False,
         class_args: List[str] = None,
         **kwargs,
     ):
@@ -34,7 +34,7 @@ class AUCCallback(LoaderMetricCallback):
             prefix: key for the metric's name
             activation: An torch.nn activation applied to the outputs.
                 Must be one of ``'none'``, ``'Sigmoid'``, or ``'Softmax'``
-            log_per_class: boolean flag to log per class metrics,
+            per_class: boolean flag to log per class metrics,
                 or use mean/macro statistics otherwise
             class_args: class names to display in the logs.
                 If None, defaults to indices for each class, starting from 0
@@ -49,7 +49,7 @@ class AUCCallback(LoaderMetricCallback):
             metric_fn=auc, activation=activation
         )
         metric_fn = wrap_class_metric2dict(
-            metric_fn, log_per_class=log_per_class, class_args=class_args
+            metric_fn, per_class=per_class, class_args=class_args
         )
         super().__init__(
             prefix=prefix,