+++ catalyst/callbacks/metrics/recall.py
@@ -17,7 +17,7 @@ class RecallCallback(BatchMetricCallback):
         output_key: str = "logits",
         prefix: str = "recall",
         activation: str = "Softmax",
-        log_per_class: bool = False,
+        per_class: bool = False,
         class_args: List[str] = None,
         **kwargs,
     ):
@@ -30,7 +30,7 @@ class RecallCallback(BatchMetricCallback):
             prefix: key for the metric's name
             activation: An torch.nn activation applied to the outputs.
                 Must be one of ``'none'``, ``'Sigmoid'``, or ``'Softmax'``
-            log_per_class: boolean flag to log per class metrics,
+            per_class: boolean flag to log per class metrics,
                 or use mean/macro statistics otherwise
             class_args: class names to display in the logs.
                 If None, defaults to indices for each class, starting from 0
@@ -45,7 +45,7 @@ class RecallCallback(BatchMetricCallback):
             metric_fn=recall, activation=activation
         )
         metric_fn = wrap_class_metric2dict(
-            metric_fn, log_per_class=log_per_class, class_args=class_args
+            metric_fn, per_class=per_class, class_args=class_args
         )
         super().__init__(
             prefix=prefix,